backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
PyJWT==2.9.0
bcrypt==4.2.0
redis==5.1.1
pymongo==4.10.1
gunicorn==22.0.0
python-dotenv==1.0.1
Pillow==10.4.0
numpy==2.1.3
transformers==4.46.3
torch==2.5.1
paddleocr==2.8.1
